move internal/operatormanifests to internal/manifests/operator
