+++ internal/manifests/operator/prometheus/prometheus-operator-alerts.yaml
